+++ tests/unit/scalyr_monitor_test.py
@@ -26,6 +26,9 @@ from scalyr_agent.scalyr_monitor import (
     MonitorConfig,
     BadMonitorConfiguration,
     define_config_option,
+    ScalyrMonitor,
+    MAX_INITIAL_GATHER_SAMPLE_SLEEP_SECS,
+    MIN_INITIAL_GATHER_SAMPLE_SLEEP_SECS,
 )
 from scalyr_agent.test_base import ScalyrTestCase
 
@@ -33,9 +36,18 @@ import scalyr_agent.util as scalyr_util
 from scalyr_agent.compat import os_environ_unicode
 
 import six
+import mock
+from six.moves import range
 
 
 class MonitorConfigTest(ScalyrTestCase):
+    @classmethod
+    def setUpClass(cls):
+        if not six.PY2:
+            super(MonitorConfigTest, cls).setUpClass()
+
+        ScalyrMonitor.DEFAULT_SAMPLE_INTERVAL_SECS = 30
+
     def test_base(self):
         test_array = ["a", 1, False]
         test_obj = {"a": 100, "b": 200}
@@ -381,6 +393,196 @@ class MonitorConfigTest(ScalyrTestCase):
             ),
             ArrayOfStrings(),
         )
+
+    def test_get_initial_sleep_delay_jitter_disabled(self):
+        logger = mock.Mock()
+
+        # 1. Default value is used
+        monitor_config = MonitorConfig(
+            {"module": "test_module"}, monitor_module="test_monitor"
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = False
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+
+        sleep_delay = monitor._get_initial_sleep_delay()
+        self.assertEqual(sleep_delay, 0)
+
+        # 2. Value is overridden on per monitor basis
+        monitor_config = MonitorConfig(
+            {"module": "test_module", "sample_interval": 10},
+            monitor_module="test_monitor",
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = False
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+
+        sleep_delay = monitor._get_initial_sleep_delay()
+        self.assertEqual(sleep_delay, 0)
+
+    def test_get_initial_sleep_delay_with_random_jitter_enabled(self):
+        logger = mock.Mock()
+
+        # 1. Default value is used
+        monitor_sample_interval = ScalyrMonitor.DEFAULT_SAMPLE_INTERVAL_SECS
+        monitor_config = MonitorConfig(
+            {"module": "test_module"}, monitor_module="test_monitor"
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = True
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+
+        for i in range(0, 100):
+            sleep_delay = monitor._get_initial_sleep_delay()
+
+            min_jitter = round(monitor_sample_interval / 10) * 2
+            max_jitter = round(monitor_sample_interval / 10) * 8
+
+            self.assertFalse(sleep_delay == monitor_sample_interval)
+            self.assertTrue(sleep_delay >= min_jitter and sleep_delay <= max_jitter)
+
+        # 2. Value is overridden on per monitor basis
+        monitor_sample_interval = 10
+        monitor_config = MonitorConfig(
+            {"module": "test_module", "sample_interval": 10},
+            monitor_module="test_monitor",
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = True
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+
+        for i in range(0, 100):
+            sleep_delay = monitor._get_initial_sleep_delay()
+
+            min_jitter = round(monitor_sample_interval / 10) * 2
+            max_jitter = round(monitor_sample_interval / 10) * 8
+
+            self.assertFalse(sleep_delay == monitor_sample_interval)
+            self.assertTrue(sleep_delay >= min_jitter and sleep_delay <= max_jitter)
+
+        # 3. Verify there is an upper bound of 40 seconds
+        monitor_sample_interval = 1000
+        monitor_config = MonitorConfig(
+            {"module": "test_module", "sample_interval": 1000},
+            monitor_module="test_monitor",
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = True
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+
+        for i in range(0, 100):
+            sleep_delay = monitor._get_initial_sleep_delay()
+            min_jitter = MIN_INITIAL_GATHER_SAMPLE_SLEEP_SECS
+            max_jitter = MAX_INITIAL_GATHER_SAMPLE_SLEEP_SECS
+
+            self.assertTrue(sleep_delay >= min_jitter and sleep_delay <= max_jitter)
+
+    def test_monitor_run_random_initial_sleep_delay_disabled(self):
+        def make_mock_thread_is_stopped(monitor):
+            def mock_is_thread_stopped(*args, **kwargs):
+                if monitor.counter >= 3:
+                    return True
+
+                monitor.counter += 1
+                return False
+
+            return mock_is_thread_stopped
+
+        mock_sleep_but_awaken_if_stopped = mock.Mock()
+
+        mock_logger = mock.Mock()
+        mock_gather_sample = mock.Mock()
+        monitor_config = MonitorConfig(
+            {"module": "test_module", "sample_interval": 10},
+            monitor_module="test_monitor",
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = False
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=mock_logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+        monitor.counter = 0
+        monitor._sleep_but_awaken_if_stopped = mock_sleep_but_awaken_if_stopped
+        monitor._is_thread_stopped = make_mock_thread_is_stopped(monitor)
+        monitor.gather_sample = mock_gather_sample
+
+        monitor.run()
+        self.assertEqual(monitor.counter, 3)
+        self.assertEqual(mock_sleep_but_awaken_if_stopped.call_count, 3)
+        self.assertEqual(mock_logger.debug.call_count, 0)
+        # 1 call is for monitor has finished message
+        self.assertEqual(mock_logger.info.call_count, 1)
+        self.assertEqual(mock_logger.exception.call_count, 0)
+
+    def test_monitor_run_random_initial_sleep_delay_enabled(self):
+        def make_mock_thread_is_stopped(monitor):
+            def mock_is_thread_stopped(*args, **kwargs):
+                if monitor.counter >= 3:
+                    return True
+
+                monitor.counter += 1
+                return False
+
+            return mock_is_thread_stopped
+
+        mock_sleep_but_awaken_if_stopped = mock.Mock()
+
+        mock_logger = mock.Mock()
+        mock_gather_sample = mock.Mock()
+        monitor_config = MonitorConfig(
+            {"module": "test_module", "sample_interval": 10},
+            monitor_module="test_monitor",
+        )
+        global_config = mock.Mock()
+        global_config.global_monitor_sample_interval_enable_jitter = True
+        monitor = ScalyrMonitor(
+            monitor_config=monitor_config,
+            logger=mock_logger,
+            sample_interval_secs=None,
+            global_config=global_config,
+        )
+        monitor.counter = 0
+        monitor._sleep_but_awaken_if_stopped = mock_sleep_but_awaken_if_stopped
+        monitor._is_thread_stopped = make_mock_thread_is_stopped(monitor)
+        monitor.gather_sample = mock_gather_sample
+
+        monitor.run()
+        self.assertEqual(monitor.counter, 3)
+        # + 1 is for initial sleep delay before first sample gather interval
+        self.assertEqual(mock_sleep_but_awaken_if_stopped.call_count, 1 + 3)
+        # 1 call is to debug function when we are initially sleeping before the gather
+        self.assertEqual(mock_logger.debug.call_count, 1)
+        # 1 call is for monitor has finished message
+        self.assertEqual(mock_logger.info.call_count, 1)
+        self.assertEqual(mock_logger.exception.call_count, 0)
 
     def get(
         self,